Rewrote the whole lib to build the maps at compile time. (#2)
Everything has been translated to the new macro form. There is a second set of tests in `revised` that should be incorporated into the existing test files, or at least renamed in the future. The docs have been mostly updated, but need one more pass. Added ability to configure some functionality, and severity of response when a key is unmapped.